Extract state utilities to kennel/state.py
Move _state_lock, load_state, save_state, clear_state, and
_resolve_git_dir from worker.py into a new kennel/state.py module.
This breaks the circular import that would result from moving sync_tasks
into tasks.py (worker imports tasks, so tasks cannot import worker).
Update events.py and tests to import from the new module.
